AI 优化项目

+++ server/core/src/main/java/ly/utils/RandomUtils.java
@@ -37,7 +37,7 @@ public class RandomUtils {
         return random.nextInt(max - min) + min;
     }
 
-    public static double RandomInt(int max) {
+    public static int RandomInt(int max) {
         return RandomInt(0, max);
     }
